+++ milvus-20231012/src/Models/GetInstanceDetailResponseBody/data/vSwitches.php
@@ -1,0 +1,62 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Milvus\V20231012\Models\GetInstanceDetailResponseBody\data;
+
+use AlibabaCloud\Dara\Model;
+
+class vSwitches extends Model
+{
+    /**
+     * @var string
+     */
+    public $vswId;
+
+    /**
+     * @var string
+     */
+    public $zoneId;
+    protected $_name = [
+        'vswId' => 'VswId',
+        'zoneId' => 'ZoneId',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->vswId) {
+            $res['VswId'] = $this->vswId;
+        }
+
+        if (null !== $this->zoneId) {
+            $res['ZoneId'] = $this->zoneId;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['VswId'])) {
+            $model->vswId = $map['VswId'];
+        }
+
+        if (isset($map['ZoneId'])) {
+            $model->zoneId = $map['ZoneId'];
+        }
+
+        return $model;
+    }
+}